requests with correct ids

+++ server/src/main/java/com/wiktor/wos/server/rest/MeetingController.java
@@ -4,11 +4,14 @@ import com.wiktor.wos.server.entity.Meeting;
 import com.wiktor.wos.server.entity.SetUpMeeting;
 import com.wiktor.wos.server.service.MeetingService;
 import com.wiktor.wos.server.service.SetUpMeetingService;
+import com.wiktor.wos.server.service.UserService;
 import com.wiktor.wos.server.service.dto.AddMeetingDTO;
 import com.wiktor.wos.server.service.dto.GuestDTO;
 import com.wiktor.wos.server.service.dto.MeetingDTO;
 import com.wiktor.wos.server.service.dto.UserDTO;
+import com.wiktor.wos.server.service.exception.ForbiddenException;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
@@ -22,11 +25,15 @@ public class MeetingController {
 
     private MeetingService meetingService;
 
+    private UserService userService;
+
     private SetUpMeetingService setUpMeetingService;
 
-    public MeetingController(MeetingService meetingService, SetUpMeetingService setUpMeetingService) {
+    public MeetingController(MeetingService meetingService, SetUpMeetingService setUpMeetingService,
+                             UserService userService) {
         this.meetingService = meetingService;
         this.setUpMeetingService = setUpMeetingService;
+        this.userService = userService;
     }
 
     @GetMapping("/{id}")
@@ -35,39 +42,37 @@ public class MeetingController {
     }
 
     @GetMapping("/user/{id}")
-    public ResponseEntity<List<MeetingDTO>> getUserPosts(@PathVariable Long id) {
+    public ResponseEntity<List<MeetingDTO>> getUserMeetings(@PathVariable Long id) {
         return ResponseEntity.ok(meetingService.getUserMeetings(id));
     }
 
     @PostMapping
-    public ResponseEntity<Void> addPost(@RequestBody @Validated AddMeetingDTO dto)
+    public ResponseEntity<Void> addMeeting(@RequestBody @Validated AddMeetingDTO dto, Authentication authentication)
             throws Exception {
-        //TODO user will be retrieved from Authentication
-        UserDTO authenticatedUser = getAuthUseDTO();
-        dto.setUserId(authenticatedUser.getId());
+        UserDTO userDTO = userService.getUser(authentication.getName());
+        dto.setUserId(userDTO.getId());
         meetingService.save(dto);
         return ResponseEntity.created(new URI("")).build();
     }
 
-    private UserDTO getAuthUseDTO() {
-        UserDTO dto = new UserDTO();
-        dto.setId(1L);
-        dto.setFirstName("user");
-        dto.setEmail("email");
-        dto.setLastName("user");
-        return dto;
-    }
-
     @PutMapping("/{id}")
-    public ResponseEntity<Void> updatePost(@RequestBody @Validated AddMeetingDTO dto, @PathVariable Long id) {
-        //TODO check if authorised user id matches meeting user id
+    public ResponseEntity<Void> updateMeeting(@RequestBody @Validated AddMeetingDTO dto, @PathVariable Long id,
+                                           Authentication authentication) {
+        UserDTO authUserDTO = userService.getUser(authentication.getName());
+        if(!authUserDTO.getId().equals(dto.getUserId())) {
+            throw new ForbiddenException("User ID does not match.");
+        }
         meetingService.update(dto, id);
         return ResponseEntity.noContent().build();
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deletePost(@PathVariable Long id) {
-        //TODO check if authorised user id matches meeting user id
+    public ResponseEntity<Void> deletePost(@PathVariable Long id, Authentication authentication) {
+        MeetingDTO meetingDTO = meetingService.getMeeting(id);
+        UserDTO authUserDTO = userService.getUser(authentication.getName());
+        if(!authUserDTO.getId().equals(meetingDTO.getUserId())) {
+            throw new ForbiddenException("User ID does not match.");
+        }
         meetingService.deletePost(id);
         return ResponseEntity.noContent().build();
     }
